feat: Controller e Router da ONG refatorado

+++ src/controllers/OngController.ts
@@ -9,15 +9,18 @@ class OngController {
       const { cnpj, nome, email, senha } = req.body;
 
       const emailVerification = await prisma.tbl_login.findUnique({
-        where: { email: email }, 
+        where: { email: email },
       });
 
       if (emailVerification) {
-        res.status(400).json({ message: `The e-mail ${email} has already been registered.` });
+        res.status(400).json({
+          message: `O email ${email} já está cadastrado.`,
+          status: 400,
+        });
         return;
       }
 
-     await prisma.tbl_login.create({
+      await prisma.tbl_login.create({
         data: {
           email: email,
           senha: senha,
@@ -37,10 +40,18 @@ class OngController {
         },
       });
       res.status(200);
-      res.json({ ResquestData: req.body, data: preCadastroOng });
+      res.json({
+        message: `O '${email}' foi registrado com sucesso.`,
+        status: 200,
+        data: preCadastroOng,
+      });
     } catch (error) {
       res.status(500);
-      res.json({ RequestData: req.body, Error: error });
+      res.json({
+        message: process.env.ERRO_500 ?? "Erro no servidor.",
+        status: 500,
+        error: error,
+      });
     }
   }
 
@@ -49,16 +60,16 @@ class OngController {
       const ong = {
         email: req.query.email.toString(),
         senha: req.query.senha.toString(),
-      }
+      };
 
       const database = await prisma.tbl_login.findUnique({
         where: { email: ong.email },
       });
 
       if (database === null) {
-        return res.status(404).json({ 
-          message: "E-mail not registered.",
-          status: false
+        return res.status(404).json({
+          message: "E-mail não foi encontrado.",
+          status: 404,
         });
       }
 
@@ -68,39 +79,58 @@ class OngController {
         });
         if (tblOng.length > 0 && tblOng.length <= 1) {
           return res.status(200).json({
-            message: "E-mail and password matched.",
-            status: true,
-            data: tblOng
+            message: "E-mail e senha conferem.",
+            status: 200,
+            data: tblOng,
           });
         }
       } else {
         return res.status(400).json({
-          message: "Password did not match.",
-          status: false
+          message: "E-mail e senha não conferem.",
+          status: 400,
         });
       }
     } catch (error) {
       console.log("Error: ", error);
+
+      return res.status(500).json({
+        message: process.env.ERRO_500 ?? "Erro no servidor.",
+        status: 500,
+        error: error,
+      });
     }
   }
 
   async getAll(req: Request, res: Response) {
-    const databaseData = await prisma.tbl_ong.findMany({});
+    const data = await prisma.tbl_ong.findMany();
 
     res.status(200);
     res.json({
-      data: databaseData,
+      message: "Todas as ONGs registradas no Banco de Dados.",
+      status: 200,
+      data: data,
     });
   }
 
   async readID(req: Request, res: Response) {
-    const data = req.params.id || req.params.idOng;
+    const data = req.params.id || req.query.id.toString();
+
     const databaseData = await prisma.tbl_ong.findUnique({
       where: { idOng: parseInt(data) },
     });
 
-    res.status(200);
-    res.json({ id: data, data: databaseData });
+    if (databaseData) {
+      res.status(200).json({ status: 200, id: data, data: databaseData });
+    } else {
+      res
+        .status(404)
+        .json({
+          message: `Não foi encontrado nenhum registro pelo id ${data}.`,
+          status: 404,
+          id: data,
+          data: databaseData,
+        });
+    }
   }
 
   async create(req: Request, res: Response) {
@@ -128,14 +158,33 @@ class OngController {
   }
 
   async remove(req: Request, res: Response) {
-    const data = req.params.id || req.params.idOng;
+    const id = req.params.id || req.query.id.toString();
 
-    const databaseData = await prisma.tbl_ong.delete({
-      where: { idOng: parseInt(data) },
+    const verification = await prisma.tbl_ong.findUnique({
+      where: {
+        idOng: Number.parseInt(id),
+      },
     });
 
-    res.status(200);
-    res.json({ RequestData: data, data: databaseData });
+    if (verification) {
+      await prisma.tbl_ong.delete({
+        where: {
+          idOng: parseInt(id),
+        },
+      });
+
+      return res.status(200).json({
+        message: `A ONG '${verification.nome}' foi excluída com sucesso.`,
+        status: 200,
+        ong: { ...verification },
+      });
+    } else {
+      return res.status(404).json({
+        message: `Nenhum registro foi encontrado pelo id '${id}'.`,
+        id: id,
+        status: 404,
+      });
+    }
   }
 }
